requirements.txt
praw
pandas
matplotlib
pyyaml
ollama